app/requirements.txt
# Main libraries
deepface==0.0.96
opencv-python==4.10.0.84
numpy==1.26.4

# Core dependencies for DeepFace (Required)
tensorflow>=2.15.0
tf-keras>=2.15.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0

# Highly recommended face detector backends
retina-face>=0.0.17
mtcnn>=0.1.1

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP + API clients
httpx>=0.27.0
openai>=1.35.0

#database
asyncpg>=0.27.0
pgvector>=0.2.4

#Whisper
faster-whisper==1.1.0
# Env loading
python-dotenv>=1.0.0

# Speaker embeddings (Resemblyzer requires torch)
torch>=2.2.0
torchaudio>=2.2.0
resemblyzer>=0.1.0
